+++ CHANGES.rst
@@ -9,6 +9,9 @@ Changes since v1.1
 - Switch to an unconditionally-stable method for the approximation of the heat equation in
   columns of the bedrock thermal layer (backward Euler time discretization instead of
   explicit time stepping).
+- Provide better error messages when trying to read in a 2D field but the input file
+  contains a 3D variable (or trying to read a 3D field but the input contains a 2D
+  variable).
 
 Changes from v1.0 to v1.1
 =========================